namespaces renamed
Renamed namespaces for better imports for the 'models' directory
