+++ server/moodle/mod/livequiz/classes/models/answer.php
@@ -21,7 +21,7 @@
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace mod_livequiz\answer;
+namespace mod_livequiz\models;
 
 use dml_exception;
