+++ server/moodle/mod/livequiz/classes/models/livequiz.php
@@ -14,10 +14,9 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace mod_livequiz\livequiz;
+namespace mod_livequiz\models;
 
 use dml_exception;
-use mod_livequiz\question\question;
 use stdClass;
 
 /**
